+++ ai/src/main/kotlin/com/oneliang/ktx/frame/ai/cnn/layer/ConvolutionLayer.kt
@@ -13,24 +13,24 @@ open class ConvolutionLayer<IN : Any, OUT : Any>(
     val filterY: Int,
     val padding: Int = 0,
     val stride: Int = 1,
-    private val forwardImpl: ((layer: ConvolutionLayer<IN, OUT>, dataId: Long, inputNeuron: IN, y: Double, training: Boolean) -> OUT)? = null,
-    private val backwardImpl: ((layer: ConvolutionLayer<IN, OUT>, dataId: Long, inputNeuron: IN, y: Double) -> Unit)? = null,
+    private val forwardImpl: ((layer: ConvolutionLayer<IN, OUT>, dataId: Long, inputNeuron: IN, y: Float, training: Boolean) -> OUT)? = null,
+    private val backwardImpl: ((layer: ConvolutionLayer<IN, OUT>, dataId: Long, inputNeuron: IN, y: Float) -> Unit)? = null,
     private val forwardResetImpl: ((layer: ConvolutionLayer<IN, OUT>, dataId: Long) -> Unit)? = null,
-    private val updateImpl: ((layer: ConvolutionLayer<IN, OUT>, epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Double) -> Unit)? = null,
+    private val updateImpl: ((layer: ConvolutionLayer<IN, OUT>, epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Float) -> Unit)? = null,
     private val initializeLayerModelDataImpl: ((layer: ConvolutionLayer<IN, OUT>, data: String) -> Unit) = { _, _ -> },
     private val saveLayerModelDataImpl: ((layer: ConvolutionLayer<IN, OUT>) -> String) = { Constants.String.BLANK }
 ) : Layer<IN, OUT>() {
 
-    var filters = Array(this.previousLayerMapDepth) { Array(this.mapDepth) { Array(this.filterY) { Array(this.filterX) { 0.0 } } } }
-    var biases = Array(this.mapDepth) { 0.0 }
+    var filters = Array(this.previousLayerMapDepth) { Array(this.mapDepth) { Array(this.filterY) { Array(this.filterX) { 0.0f } } } }
+    var biases = Array(this.mapDepth) { 0.0f }
     val outX = calculateOutSize(inX, padding, filterX, stride)
     val outY = calculateOutSize(inY, padding, filterY, stride)
 
-    override fun forwardImpl(dataId: Long, inputNeuron: IN, y: Double, training: Boolean): OUT {
+    override fun forwardImpl(dataId: Long, inputNeuron: IN, y: Float, training: Boolean): OUT {
         return this.forwardImpl?.invoke(this, dataId, inputNeuron, y, training) ?: outputNullError()
     }
 
-    override fun backwardImpl(dataId: Long, inputNeuron: IN, y: Double) {
+    override fun backwardImpl(dataId: Long, inputNeuron: IN, y: Float) {
         this.backwardImpl?.invoke(this, dataId, inputNeuron, y)
     }
 
@@ -38,7 +38,7 @@ open class ConvolutionLayer<IN : Any, OUT : Any>(
         this.forwardResetImpl?.invoke(this, dataId)
     }
 
-    override fun updateImpl(epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Double) {
+    override fun updateImpl(epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Float) {
         this.updateImpl?.invoke(this, epoch, printPeriod, totalDataSize, learningRate)
     }
