+++ ai/src/main/kotlin/com/oneliang/ktx/frame/ai/cnn/layer/SoftmaxLayer.kt
@@ -8,22 +8,22 @@ import com.oneliang.ktx.util.concurrent.atomic.AtomicMap
 open class SoftmaxLayer<IN : Any, OUT : Any, LOSS : Any>(
     val neuronCount: Int,
     val typeCount: Int,
-    private val forwardImpl: ((layer: SoftmaxLayer<IN, OUT, LOSS>, dataId: Long, inputNeuron: IN, y: Double, training: Boolean) -> OUT)? = null,
-    private val backwardImpl: ((layer: SoftmaxLayer<IN, OUT, LOSS>, dataId: Long, inputNeuron: IN, y: Double) -> Unit)? = null,
+    private val forwardImpl: ((layer: SoftmaxLayer<IN, OUT, LOSS>, dataId: Long, inputNeuron: IN, y: Float, training: Boolean) -> OUT)? = null,
+    private val backwardImpl: ((layer: SoftmaxLayer<IN, OUT, LOSS>, dataId: Long, inputNeuron: IN, y: Float) -> Unit)? = null,
     private val forwardResetImpl: ((layer: SoftmaxLayer<IN, OUT, LOSS>, dataId: Long) -> Unit)? = null,
-    private val updateImpl: ((layer: SoftmaxLayer<IN, OUT, LOSS>, epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Double) -> Unit)? = null,
+    private val updateImpl: ((layer: SoftmaxLayer<IN, OUT, LOSS>, epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Float) -> Unit)? = null,
     private val initializeLayerModelDataImpl: ((layer: SoftmaxLayer<IN, OUT, LOSS>, data: String) -> Unit) = { _, _ -> },
     private val saveLayerModelDataImpl: ((layer: SoftmaxLayer<IN, OUT, LOSS>) -> String) = { Constants.String.BLANK }
 ) : LossLayer<IN, OUT, LOSS>() {
 
-    var derivedWeights = AtomicMap<String, Array<Array<Double>>>()//: Array<Array<Double>> = Array(neuronCount) { Array(this.typeCount) { 0.0 } }
-    var weights: Array<Array<Double>> = emptyArray()
+    var derivedWeights = AtomicMap<String, Array<Array<Float>>>()//: Array<Array<Float>> = Array(neuronCount) { Array(this.typeCount) { 0.0 } }
+    var weights: Array<Array<Float>> = emptyArray()
 
-    override fun forwardImpl(dataId: Long, inputNeuron: IN, y: Double, training: Boolean): OUT {
+    override fun forwardImpl(dataId: Long, inputNeuron: IN, y: Float, training: Boolean): OUT {
         return this.forwardImpl?.invoke(this, dataId, inputNeuron, y, training) ?: outputNullError()
     }
 
-    override fun backwardImpl(dataId: Long, inputNeuron: IN, y: Double) {
+    override fun backwardImpl(dataId: Long, inputNeuron: IN, y: Float) {
         this.backwardImpl?.invoke(this, dataId, inputNeuron, y)
     }
 
@@ -31,7 +31,7 @@ open class SoftmaxLayer<IN : Any, OUT : Any, LOSS : Any>(
         this.forwardResetImpl?.invoke(this, dataId)
     }
 
-    override fun updateImpl(epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Double) {
+    override fun updateImpl(epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Float) {
         this.updateImpl?.invoke(this, epoch, printPeriod, totalDataSize, learningRate)
     }
