+++ ai/src/main/kotlin/com/oneliang/ktx/frame/ai/cnn/layer/TransformLayer.kt
@@ -4,19 +4,19 @@ import com.oneliang.ktx.Constants
 import com.oneliang.ktx.frame.ai.dnn.layer.LossLayer
 
 open class TransformLayer<IN : Any, OUT : Any, LOSS : Any>(
-    private val forwardImpl: ((layer: TransformLayer<IN, OUT, LOSS>, dataId: Long, inputNeuron: IN, y: Double, training: Boolean) -> OUT)? = null,
-    private val backwardImpl: ((layer: TransformLayer<IN, OUT, LOSS>, dataId: Long, inputNeuron: IN, y: Double) -> Unit)? = null,
+    private val forwardImpl: ((layer: TransformLayer<IN, OUT, LOSS>, dataId: Long, inputNeuron: IN, y: Float, training: Boolean) -> OUT)? = null,
+    private val backwardImpl: ((layer: TransformLayer<IN, OUT, LOSS>, dataId: Long, inputNeuron: IN, y: Float) -> Unit)? = null,
     private val forwardResetImpl: ((layer: TransformLayer<IN, OUT, LOSS>, dataId: Long) -> Unit)? = null,
-    private val updateImpl: ((layer: TransformLayer<IN, OUT, LOSS>, epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Double) -> Unit)? = null,
+    private val updateImpl: ((layer: TransformLayer<IN, OUT, LOSS>, epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Float) -> Unit)? = null,
     private val initializeLayerModelDataImpl: ((layer: TransformLayer<IN, OUT, LOSS>, data: String) -> Unit) = { _, _ -> },
     private val saveLayerModelDataImpl: ((layer: TransformLayer<IN, OUT, LOSS>) -> String) = { Constants.String.BLANK }
 ) : LossLayer<IN, OUT, LOSS>() {
 
-    override fun forwardImpl(dataId: Long, inputNeuron: IN, y: Double, training: Boolean): OUT {
+    override fun forwardImpl(dataId: Long, inputNeuron: IN, y: Float, training: Boolean): OUT {
         return this.forwardImpl?.invoke(this, dataId, inputNeuron, y, training) ?: outputNullError()
     }
 
-    override fun backwardImpl(dataId: Long, inputNeuron: IN, y: Double) {
+    override fun backwardImpl(dataId: Long, inputNeuron: IN, y: Float) {
         this.backwardImpl?.invoke(this, dataId, inputNeuron, y)
     }
 
@@ -24,7 +24,7 @@ open class TransformLayer<IN : Any, OUT : Any, LOSS : Any>(
         this.forwardResetImpl?.invoke(this, dataId)
     }
 
-    override fun updateImpl(epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Double) {
+    override fun updateImpl(epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Float) {
         this.updateImpl?.invoke(this, epoch, printPeriod, totalDataSize, learningRate)
     }
